+++ EnhanceQoL/NewSettingsTable.lua
@@ -5,6 +5,7 @@ addon.variables.NewVersionTableEQOL = {
 	["EQOL_Quest"] = true,
 	["EQOL_Social"] = true,
 	["EQOL_autoAcceptSummon"] = true,
+	["EQOL_autoAcceptResurrection"] = true,
 	["EQOL_autoReleasePvP"] = true,
 	["EQOL_quickSkipCinematic"] = true,
 }